requirements.txt
PyQt5
pynput
pyatspi2
